All tables are finished ("no data" handling, correct titles)

+++ src/main/java/comp3111/covid/TableEntry.java
@@ -2,10 +2,10 @@ package comp3111.covid;
 
 public class TableEntry {
 	String name;
-	Integer data1;
-	Integer data2;
+	Float data1;
+	Float data2;
 	
-	TableEntry(String name, Integer data1, Integer data2){
+	TableEntry(String name, Float data1, Float data2){
 		this.name=name;
 		this.data1=data1;
 		this.data2=data2;
@@ -15,11 +15,11 @@ public class TableEntry {
 		return name;
 	}
 	
-	public Integer getData1() {
+	public Float getData1() {
 		return data1;
 	}
 	
-	public Integer getData2() {
+	public Float getData2() {
 		return data2;
 	}
 }
